audio attempt

+++ Assets/Scripts/Utils/SpinRandom.cs
@@ -4,9 +4,10 @@ namespace Game
 {
     public class SpinRandom : MonoBehaviour
     {
-        public int SpeedMin;
-        public int SpeedMax;
+        public float SpeedMin;
+        public float SpeedMax;
 
+        //private void Update() => transform.Rotate(0f, 0f, Random.Range(SpeedMin, SpeedMax) * Time.deltaTime, Space.Self);
         private void Update() => transform.Rotate(0f, 0f, Random.Range(SpeedMin, SpeedMax) * Time.deltaTime, Space.Self);
     }
 }
